+++ tests/lammps_tests/test_mliap.py
@@ -193,7 +193,7 @@ def _lammps_results_to_atoms(lammps_log, force_dump):
     return latoms
 
 
-def _run_mliap(atoms, pair_style, wd, command, test_name):
+def _run_mliap(atoms, pair_style, wd, command, test_name, timeout=30):
     wd = wd.resolve()
     pbc = atoms.get_pbc()
     pbc_str = ' '.join(['p' if x else 'f' for x in pbc])
@@ -231,7 +231,7 @@ def _run_mliap(atoms, pair_style, wd, command, test_name):
         f'{command} -in {input_script_path} '
         f'-k on g 1 -sf kk -pk kokkos neigh half -log {lammps_log}'
     )
-    subprocess_routine(command.split(), test_name)
+    subprocess_routine(command.split(), test_name, timeout=timeout)
 
     lmp_atoms = _lammps_results_to_atoms(lammps_log, force_dump)
     assert lmp_atoms.calc is not None
@@ -253,15 +253,15 @@ def _run_mliap(atoms, pair_style, wd, command, test_name):
 
 
 def mliap_lammps_run(
-    atoms, pt_path, wd, test_name, lammps_cmd
+    atoms, pt_path, wd, test_name, lammps_cmd, timeout=30
 ):
     command = lammps_cmd
     pair_style = f'{pt_path}'
-    return _run_mliap(atoms, pair_style, wd, command, test_name)
+    return _run_mliap(atoms, pair_style, wd, command, test_name, timeout=timeout)
 
 
-def subprocess_routine(cmd, name):
-    res = subprocess.run(cmd, capture_output=True, timeout=30)
+def subprocess_routine(cmd, name, timeout=30):
+    res = subprocess.run(cmd, capture_output=True, timeout=timeout)
     if res.returncode != 0:
         logger.error(f'Subprocess {name} failed return code: {res.returncode}')
         logger.error('==== STDERR ====')
@@ -330,6 +330,7 @@ def test_mliap_cueq(
         wd=tmp_path,
         test_name='mliap cueq',
         lammps_cmd=lammps_cmd,
+        timeout=120,
     )
     atoms.calc = ref_7net0_calculator
     assert_atoms(atoms, atoms_lmp, atol=1e-5)